Refactor Task, moved to Task package
