+++ src/main/java/com/organizer/model/task/Task.java
@@ -1,4 +1,4 @@
-package com.organizer.model;
+package com.organizer.model.task;
 
 public class Task {
